Fix APRS UDP issue
Common APRS-IS require authentication in each UDP frame by specs (http://www.aprs-is.net/SendOnlyPorts.aspx).
It is missing and APRS does not work with usual APRS2 IS servers. Maybe this is the reason why the last updates since moving to UDP are not used. I wonder why nobody claims about it.